# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvalCaseOut": {
        "description": "An eval case conforming to the frozen eval-case format (#8, ADR-0019):\nan input prompt plus the grader that judges the answer. Emitted by the\npromote-a-trace-to-an-eval-case endpoint (#259).\n\n``shared_history`` mirrors the worker's ``EvalCase`` field (#550, ADR-0051):\na promoted trace is a standalone case, so it emits the ``False`` default\n(fresh conversation). Kept here to satisfy the schema field-parity gate; the\npromote endpoint has no reason to mint a history-chained case.\n\n``expect_status`` mirrors the frozen ``ExpectedStatus`` (#262, ADR-0053): the\nterminal session status the case asserts, default ``done``. A promoted trace\nis a completed conversation, so the emitted case keeps the default; a human\nedits it to ``awaiting-approval`` when the case should assert an approval gate\nheld. Do not let this literal drift from the schema's ``ExpectedStatus`` enum.",
        "properties": {
          "expect_status": {
            "default": "done",
            "enum": [
              "done",
              "awaiting-approval"
            ],
            "title": "Expect Status",
            "type": "string"
          },
          "grader": {
            "$ref": "#/components/schemas/GraderOut"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "shared_history": {
            "default": false,
            "title": "Shared History",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "input",
          "grader"
        ],
        "title": "EvalCaseOut",
        "type": "object"
      },
...
      "GraderOut": {
        "description": "A deterministic grader, mirroring the frozen eval-case Grader shape\n(`apps/worker/schema/eval-cases.schema.json`). Do not let this drift from the\nworker's `Grader` model.",
        "properties": {
          "case_sensitive": {
            "default": false,
            "title": "Case Sensitive",
            "type": "boolean"
          },
          "expected": {
            "title": "Expected",
            "type": "string"
          },
          "kind": {
            "enum": [
              "exact",
              "contains",
              "regex",
              "tool_called"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "expected"
        ],
        "title": "GraderOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationNode": {
        "description": "One node in the reconstructed observation tree (Runs view).",
        "properties": {
          "children": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ObservationNode"
            },
            "title": "Children",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "startTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starttime"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "usageDetails": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usagedetails"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "title": "ObservationNode",
        "type": "object"
      },
      "PodLogs": {
        "description": "Runner-pod logs for the per-run runner-logs affordance.",
        "properties": {
          "container": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container"
          },
          "logs": {
            "title": "Logs",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "pod": {
            "title": "Pod",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "pod",
          "container",
          "logs"
        ],
        "title": "PodLogs",
        "type": "object"
      },
      "RunnerPods": {
        "description": "The runner sandbox pods in a namespace (populates the Logs pod dropdown).",
        "properties": {
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "pods": {
            "items": {
              "type": "string"
            },
            "title": "Pods",
            "type": "array"
          }
        },
        "required": [
          "namespace",
          "pods"
        ],
        "title": "RunnerPods",
        "type": "object"
      },
...
      "TraceTree": {
        "description": "A trace plus its reconstructed observation tree.",
        "properties": {
          "approval_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Decision"
          },
          "sandbox_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Id"
          },
          "trace": {
            "additionalProperties": true,
            "title": "Trace",
            "type": "object"
          },
          "tree": {
            "items": {
              "$ref": "#/components/schemas/ObservationNode"
            },
            "title": "Tree",
            "type": "array"
          }
        },
        "required": [
          "trace",
          "tree"
        ],
        "title": "TraceTree",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Curie API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/langfuse/traces": {
      "get": {
        "operationId": "list_traces_langfuse_traces_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Traces Langfuse Traces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Traces",
        "tags": [
          "langfuse"
        ]
      }
    },
    "/langfuse/traces/{trace_id}": {
      "get": {
        "operationId": "get_trace_langfuse_traces__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceTree"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trace",
        "tags": [
          "langfuse"
        ]
      }
    },
    "/langfuse/traces/{trace_id}/eval-case": {
      "post": {
        "description": "Promote a real trace into an anonymized, runnable eval case (#259).\n\nReads the trace + its observations, extracts the conversation input/output,\nanonymizes them (emails, Slack ids, phone numbers, credential-shaped tokens),\nand emits an EvalCase conforming to the frozen eval-case format. 404s when the\ntrace has no observations yet, matching the read path.",
        "operationId": "promote_trace_to_eval_case_langfuse_traces__trace_id__eval_case_post",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalCaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Trace To Eval Case",
        "tags": [
          "langfuse"
        ]
      }
    },
...
    "/observability/runners": {
      "get": {
        "operationId": "list_runner_pods_observability_runners_get",
        "parameters": [
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerPods"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runner Pods",
        "tags": [
          "observability"
        ]
      }
    },
    "/observability/runners/{namespace}/{pod}/logs": {
      "get": {
        "operationId": "runner_logs_observability_runners__namespace___pod__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pod",
            "required": true,
            "schema": {
              "title": "Pod",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "container",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Container"
            }
          },
          {
            "in": "query",
            "name": "tail_lines",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tail Lines"
            }
          },
          {
            "in": "query",
            "name": "previous",
            "required": false,
            "schema": {
              "default": false,
              "title": "Previous",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodLogs"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runner Logs",
        "tags": [
          "observability"
        ]
      }
    }
  }
}